backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-cloud-vision==3.4.4
supabase==1.0.4
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.0